# Replace global Mutex<Hashmap> with DashMap