This patch does the following:
* removes the global mutex which heavily impacts multi-threading
* removes dropping of interned strings

Note for swc: 

swc need to add back the dropping part otherwise applications may blow up their memory.

Impact on real project:

* Bottleneck on the global mutex is removed so we see a 10s -> 1s perf gain around string-cache.
* A full rspack build of an internal project reduced from 4.4s -> 3.2s (~30%)

![image](https://user-images.githubusercontent.com/1430279/218393076-bc4eec8c-d0a5-46a2-a7b6-3fa974628570.png)