@Boshen 2023-02-16
This approach does not work. Using hash as the key will result in hash collision, using `(string, hash)` as the key will result in an string extra allocation.